ai-agent/requirements.txt
langgraph
langchain-google-genai
nats-py
python-dotenv
pydantic
httpx

